requirements.txt
flask
requests
python-dotenv
msal
psycopg2-binary
